Begin splitting out unit cards
