+++ src/factories/cards/makeCards.ts
@@ -1,8 +1,15 @@
 import cloneDeep from 'lodash.clonedeep';
 import { v4 as uuidv4 } from 'uuid';
 
-import { Card, CardType, ResourceCard } from '@/types/cards';
+import {
+    Card,
+    CardType,
+    ResourceCard,
+    UnitBase,
+    UnitCard,
+} from '@/types/cards';
 import { Resource } from '@/types/resources';
+import { PassiveEffect } from '@/types/effects';
 
 export const makeResourceCard = (resource: Resource): ResourceCard => ({
     cardType: CardType.RESOURCE,
@@ -11,6 +18,20 @@ export const makeResourceCard = (resource: Resource): ResourceCard => ({
     isUsed: false,
     id: uuidv4(),
 });
+
+export const makeUnitCard = (unitBase: UnitBase): UnitCard => {
+    const hasQuick = unitBase.passiveEffects.indexOf(PassiveEffect.QUICK) > -1;
+    return {
+        ...unitBase,
+        cardType: CardType.UNIT,
+        hp: unitBase.totalHp,
+        numAttacksLeft: hasQuick ? unitBase.numAttacks : 0,
+        isSelected: false,
+        hpBuff: 0,
+        attackBuff: 0,
+        originalCost: cloneDeep(unitBase.cost),
+    };
+};
 
 export const makeCard = <T = Card>(card: T): T => {
     return { ...cloneDeep(card), id: uuidv4() };
